refactor(backend): reorganize
