+++ packages/apps/backend/src/apis/app/users/users.controller.ts
@@ -1,10 +1,11 @@
 import { Body, Delete, Get, HttpCode, HttpStatus, Param, ParseUUIDPipe, Post, Put, UseGuards } from '@nestjs/common';
 import { ApiBody, ApiOperation, ApiParam, ApiResponse, ApiTags } from '@nestjs/swagger';
 
+import { UserAuthGuard } from '../utils/guards/user-auth.guard';
+
+import { CurrentUser, type CurrentUserData } from '@/apis/app/utils/decorators/current-user.decorator'
 import { ApiStandardResponses } from '@/apis/utils/api-default-response';
 import { ApiController } from '@/apis/utils/controllers/api.controller';
-import { CurrentUser, type CurrentUserData } from '@/apis/utils/decorators/current-user.decorator';
-import { UserAuthGuard } from '@/apis/utils/guards/user-auth.guard';
 import { UserCommandService } from '@/domain/aggregates/user/user.command.service';
 import { UserQueryService } from '@/domain/aggregates/user/user.query.service';
 import { CreateUserInputDto, UpdateUserDataDto, UserResponseDto } from '@/domain/aggregates/user/utils/dto';
